user login and logout controller added

+++ src/utils/cloudinary.js
@@ -1,31 +1,34 @@
-import { v2 as cloudinary } from 'cloudinary';
-import fs from 'fs';
+import { v2 as cloudinary } from "cloudinary";
+import fs from "fs";
 
-    // Configuration
-    cloudinary.config({ 
-        cloud_name: process.env.CLOUDINARY_CLOUD_NAME, 
-        api_key: process.env.CLOUDINARY_API_KEY, 
-        api_secret: process.env.CLOUDINARY_API_SECRET
+// Configuration
+cloudinary.config({
+  cloud_name: process.env.CLOUDINARY_CLOUD_NAME,
+  api_key: process.env.CLOUDINARY_API_KEY,
+  api_secret: process.env.CLOUDINARY_API_SECRET,
+});
+
+const uploadOnCloudinary = async (localFilePath) => {
+  try {
+    if (!localFilePath) return null;
+
+    const response = await cloudinary.uploader.upload(localFilePath, {
+      resource_type: "auto",
     });
-    
-    const uploadOnCloudinary = async (localFilePath) => {
-        try {
-            if (!localFilePath) return null
 
-            //upload the file on cloudinary
-            const response = await cloudinary.uploader.upload(localFilePath, {
-                 resource_type: "auto"
-            })
+    // remove local file after successful upload
+    fs.existsSync(localFilePath) && fs.unlinkSync(localFilePath);
 
-            // file has been uploaded successfully on cloudinary
-            // console.log("File is uploaded on clodinary", response.url);
-            fs.unlinkSync(localFilePath)
-            return response;
-
-        } catch (error) {
-            fs.unlinkSync(localFilePath)  // remove the locally saved temporary file as the upload operation got failed
-            return null;
-        }
+    return response;
+  } catch (error) {
+    // cleanup if upload fails
+    if (fs.existsSync(localFilePath)) {
+      fs.unlinkSync(localFilePath);
     }
 
-export {uploadOnCloudinary}
+    console.error("Cloudinary upload failed:", error);
+    return null;
+  }
+};
+
+export { uploadOnCloudinary };
